+++ beacon_chain/validators/message_router.nim
@@ -121,7 +121,7 @@ proc routeSignedBeaconBlock*(
       signature = shortLog(blck.signature), error = res.error()
 
   let newBlockRef = await router[].blockProcessor.storeBlock(
-    MsgSource.api, sendTime, blck, BlobSidecars @[])
+    MsgSource.api, sendTime, blck, Opt.none(BlobSidecars))
 
   # The boolean we return tells the caller whether the block was integrated
   # into the chain