+++ tests/testblockutil.nim
@@ -520,11 +520,13 @@ iterator makeTestBlocks*(
   state: ForkedHashedBeaconState,
   cache: var StateCache,
   blocks: int,
-  attested: bool,
+  eth1_data = Eth1Data(),
+  attested = false,
+  allDeposits = newSeq[Deposit](),
   syncCommitteeRatio = 0.0,
+  graffiti = default(GraffitiBytes),
   cfg = defaultRuntimeConfig): ForkedSignedBeaconBlock =
-  var
-    state = assignClone(state)
+  var state = assignClone(state)
   for _ in 0..<blocks:
     let
       parent_root = withState(state[]): forkyState.latest_block_root
@@ -534,7 +536,24 @@ iterator makeTestBlocks*(
             state[], parent_root, getStateField(state[], slot), cache)
         else:
           @[]
+      stateEth1 = getStateField(state[], eth1_data)
+      stateDepositIndex = getStateField(state[], eth1_deposit_index)
+      deposits =
+        if stateDepositIndex < stateEth1.deposit_count:
+          let
+            lowIndex = stateDepositIndex
+            numDeposits = min(MAX_DEPOSITS, stateEth1.deposit_count - lowIndex)
+            highIndex = lowIndex + numDeposits - 1
+          allDeposits[lowIndex .. highIndex]
+        else:
+          newSeq[Deposit]()
       sync_aggregate = makeSyncAggregate(state[], syncCommitteeRatio, cfg)
 
-    yield addTestBlock(state[], cache,
-      attestations = attestations, sync_aggregate = sync_aggregate, cfg = cfg)
+    yield addTestBlock(
+      state[], cache,
+      eth1_data = eth1_data,
+      attestations = attestations,
+      deposits = deposits,
+      sync_aggregate = sync_aggregate,
+      graffiti = graffiti,
+      cfg = cfg)